Remove redundant CountDownLatch
We can use InstantTaskExecutorRule to run all observers synchronous.
An error will be thrown when a test exceeds more than 4 seconds. The
value is increased to 4 seconds comparing to the 2 seconds in
LiveDataUtil because the 4 seconds are responsible for a whole test
method rather than a single call to getValue(). This also removes a
potential bug while testing on really low devices where an observer
could need more than 2 seconds to set its value which causes the latch
to remove its lock and so introduce unsafe threads.