move coroutines to the new testing artifact
This CL updates the coroutines testing to a new artifact and
also introduces a wrapper for all app executors.

Due to the Room's transaction handling, we cannot run single
threaded tests w/ transactions so for room we have a separate
executor. It is also not a blocking idling resource because
blocking the test thread might block the transaction's execution
so we do a busy loop idle check for it :/.