Prevent memory leaks in AutoClearedValue
When using AutoClearedValue we expect it to clear the value as soon
as the fragment view is destroyed. But in fact, the value is cleared
only when the fragment itself is destroyed, which makes AutoClearedValue
useless as the destroyed fragment is already available to be garbage
collected.

In order to solve the problem we should instead use viewLifecycleOwner.
But it's not available right away, so we wait till fragment lifecycle
ON_START event.

We also fixed AutoClearedValueTest so its TestFragment returns a non-null
View, which prevents IllegalStateException.