Avoid references to the Fragment's Views after onDestroyView()
After onDestroyView() runs, Fragment's Views are eligible for garbage collection. Ensure that we don't hold onto any references after onDestroyView() to allow for the Views to be garbage collected as soon as possible.

For the GithubBrowserSample, this required changing the AutoClearedValue class to clear the backing field when the Fragment's view is destroyed, rather than wait for the Fragment itself to be destroyed.

Test: ./run_all_tests.sh